Make GetReplicas return int32

+++ pkg/controller/experiment/converter.go
@@ -37,7 +37,7 @@ func ConvertExperiment(in *redskyv1alpha1.Experiment, out *redskyapi.Experiment)
 	if in.Spec.Parallelism != nil {
 		out.Optimization.ParallelTrials = *in.Spec.Parallelism
 	} else {
-		out.Optimization.ParallelTrials = int32(in.GetReplicas())
+		out.Optimization.ParallelTrials = in.GetReplicas()
 	}
 	if in.Spec.BurnIn != nil {
 		out.Optimization.BurnIn = *in.Spec.BurnIn
